+++ src/main/java/csc1035/project2/RoomHandler.java
@@ -158,6 +158,11 @@ public class RoomHandler {
      */
     public void cancelReservation ( String id ) {
         IController ic = new Controller();
+        Booking b = (Booking) ic.getById( Booking.class, id );
+
+        if(b.getStaffBooking() != null) {ic.delete( StaffBooking.class, id );}
+        else if(b.getStudentBooking() != null) {ic.delete( StudentBooking.class, id );}
+        else if(b.getModuleBooking() != null) {ic.delete( ModuleBooking.class, id );}
         ic.delete( Booking.class, id );
         refreshRoomHandler();
     }